FAST_API/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# Core (CPU)
torch==2.8.0+cpu
torchvision==0.23.0+cpu

# YOLO/비전
ultralytics>=8.3.155   # numpy 2.x와 호환되는 버전권장
opencv-python-headless==4.10.* 
fashion-clip==0.2.2

# FastAPI
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.2
pydantic==2.11.7
pydantic_core==2.33.2

# DB/ORM
SQLAlchemy==1.4.54
psycopg2==2.9.10
qdrant-client==1.8.0

# OpenAI
openai==1.97.1

# AWS
boto3==1.40.2
botocore==1.40.2

# Data
pandas==2.3.1
numpy>=2.0,<3.0
 
# HTTP
httpx==0.28.1
httpcore==1.0.9
requests==2.32.5

# Auth/Sec
Authlib==1.6.1
cryptography==45.0.5
bcrypt==4.3.0
PyJWT==2.10.1

# Utils
python-dotenv==1.1.1
python-multipart==0.0.20
Jinja2==3.1.6
Pillow==10.4.0
itsdangerous==2.2.0

# Text / Fuzzy
rapidfuzz==3.5.2

# Monitoring
prometheus-client==0.19.0